# openapi2-functions.yaml
swagger: '2.0'
info:
  title: chattalker backend
  description: API of chat message talker on Twitch for OBS and so on
  version: 0.1.0
schemes:
  - https
produces:
  - application/json
securityDefinitions:
  firebase:
    authorizationUrl: ''
    flow: 'implicit'
    type: 'oauth2'
    x-google-issuer: 'https://securetoken.google.com/chattalker'
    x-google-jwks_uri: 'https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com'
    x-google-audiences: 'chattalker'
paths:
  /text-to-speech:
    get:
      summary: Text-to-speech
      operationId: text-to-speech
      x-google-backend:
        address: https://text-to-speech-bf7bhumxka-uc.a.run.app
      produces:
        - multipart/form-data
      security:
        - firebase: []
      parameters:
        - in: query
          name: text
          type: string
        - in: query
          name: 'voice[en]'
          type: string
        - in: query
          name: 'voice[ja]'
          type: string
        - in: query
          name: 'voice[und]'
          type: string
      responses:
        '200':
          description: A successful response
          schema:
            type: object
            properties:
              audioContent:
                type: string
              language:
                type: string
        '204':
          description: Keep Alive
    options:
      summary: Text-to-speech Preflight
      operationId: text-to-speech-preflight
      x-google-backend:
        address: https://text-to-speech-bf7bhumxka-uc.a.run.app
      responses:
        '204':
          description: A successful response

  /youtube-oauth2callback:
    get:
      summary: YouTube OAuth2 Callback
      operationId: youtube-oauth2callback
      x-google-backend:
        address: https://youtube-oauth2callback-bf7bhumxka-uc.a.run.app
      security:
        - firebase: []
      parameters:
        - in: query
          name: code
          type: string
        - in: query
          name: redirectUri
          type: string
      responses:
        '200':
          description: A successful response
          schema:
            type: object
            properties:
              access_token:
                type: string
              expires_in:
                type: integer
              refresh_token:
                type: string
              scope:
                type: string
              token_type:
                type: string

    options:
      summary: YouTube OAuth2 Callback Preflight
      operationId: youtube-oauth2callback-preflight
      x-google-backend:
        address: https://youtube-oauth2callback-bf7bhumxka-uc.a.run.app
      responses:
        '204':
          description: A successful response

  /youtube-oauth2refresh:
    get:
      summary: YouTube OAuth2 Refresh
      operationId: youtube-oauth2refresh
      x-google-backend:
        address: https://youtube-oauth2refresh-bf7bhumxka-uc.a.run.app
      security:
        - firebase: []
      parameters:
        - in: query
          name: refreshToken
          type: string
      responses:
        '200':
          description: A successful response
          schema:
            type: object
            properties:
              access_token:
                type: string
              expires_in:
                type: integer
              scope:
                type: string
              token_type:
                type: string
    options:
      summary: YouTube OAuth2 Refresh Preflight
      operationId: youtube-oauth2refresh-preflight
      x-google-backend:
        address: https://youtube-oauth2refresh-bf7bhumxka-uc.a.run.app
      responses:
        '204':
          description: A successful response

  /authenticate-with-token:
    get:
      summary: Authenticate with token
      operationId: authenticate-with-token
      x-google-backend:
        address: https://authenticate-with-token-bf7bhumxka-uc.a.run.app
      parameters:
        - in: query
          name: token
          type: string
        - in: query
          name: uid
          type: string
      responses:
        '200':
          description: A successful response
          schema:
            type: string